feat(sparkle): replace ContentMessage with new design
- New variants: error/success/info/warning/gray (replaces primary/highlight/green/blue/rose/golden/outline)
- New sizes: sm/xs (replaces sm/md/lg)
- New layout: icon+title row with body stacked below; inline icon when no title
- Dark mode: semantic tokens (success, highlight) auto-remap; raw tokens (red, orange, gray) use dark: prefix
- Remove ContentMessageInline (merged into ContentMessage layout)
- Remove NewContentMessage (design merged into ContentMessage)
- Update stories with new variants/sizes
