+++ sparkle/src/components/ContentMessage.tsx
@@ -5,114 +5,51 @@ import { cva } from "class-variance-authority";
 import React, { type ComponentType } from "react";
 
 const CONTENT_MESSAGE_VARIANTS = [
-  "primary",
-  "warning",
+  "error",
   "success",
-  "highlight",
   "info",
-  "green",
-  "blue",
-  "rose",
-  "golden",
-  "outline",
+  "warning",
+  "gray",
 ] as const;
 
 type ContentMessageVariantType = (typeof CONTENT_MESSAGE_VARIANTS)[number];
 
-const CONTENT_MESSAGE_SIZES = ["sm", "md", "lg"] as const;
+const CONTENT_MESSAGE_SIZES = ["sm", "xs"] as const;
 
 type ContentMessageSizeType = (typeof CONTENT_MESSAGE_SIZES)[number];
 
-const sharedVariantStyles = {
-  primary: "bg-muted-background",
-  success: "bg-success-100",
-  warning: "bg-warning-100",
-  highlight: "bg-highlight-100",
-  info: "bg-info-100",
-  green: "bg-success-100",
-  blue: "bg-highlight-100",
-  rose: "bg-warning-100",
-  golden: "bg-info-100",
-  outline: "bg-transparent border border-border",
-};
-
-const contentMessageVariants = cva(
-  "flex flex-col gap-1 rounded-2xl p-4 pl-5 min-h-[52px]",
-  {
-    variants: {
-      variant: sharedVariantStyles,
-      size: {
-        lg: "",
-        md: "max-w-xl",
-        sm: "max-w-sm",
-      },
-    },
-    defaultVariants: {
-      variant: "info",
-      size: "md",
-    },
-  }
-);
-
-const contentMessageInlineVariants = cva(
-  "flex items-center gap-3 rounded-xl p-3 pl-4 min-h-[52px]",
-  {
-    variants: {
-      variant: sharedVariantStyles,
-    },
-    defaultVariants: {
-      variant: "info",
-    },
-  }
-);
-
-const iconVariants = cva("shrink-0", {
+// success-100 and highlight-100 are semantic tokens that remap automatically in
+// dark mode (tokens.css .dark {}). red, orange, gray have no semantic alias so
+// we pair them with explicit dark: variants following the same inversion pattern
+// (100 → 900 for bg, 800 → 200 for text).
+const containerVariants = cva("flex flex-col", {
   variants: {
     variant: {
-      primary: "text-muted-foreground",
-      warning: "text-warning-900",
-      success: "text-success-900",
-      highlight: "text-highlight-900",
-      info: "text-info-900",
-      green: "text-success-900",
-      blue: "text-highlight-900",
-      rose: "text-warning-900",
-      golden: "text-info-900",
-      outline: "text-muted-foreground",
+      error: "bg-red-100 dark:bg-red-900",
+      success: "bg-success-100",
+      info: "bg-highlight-100",
+      warning: "bg-orange-100 dark:bg-orange-900",
+      gray: "bg-gray-100 dark:bg-gray-900",
     },
+    size: {
+      sm: "gap-3 p-4 rounded-2xl",
+      xs: "gap-3 p-3 rounded-xl",
+    },
+  },
+  defaultVariants: {
+    variant: "info",
+    size: "sm",
   },
 });
 
-const titleVariants = cva("heading-sm", {
+const colorVariants = cva("", {
   variants: {
     variant: {
-      primary: "text-foreground",
-      warning: "text-warning-900",
-      success: "text-success-900",
-      highlight: "text-highlight-900",
-      info: "text-info-900",
-      green: "text-success-900",
-      blue: "text-highlight-900",
-      rose: "text-warning-900",
-      golden: "text-info-900",
-      outline: "text-foreground",
-    },
-  },
-});
-
-const textVariants = cva("text-sm", {
-  variants: {
-    variant: {
-      primary: "text-muted-foreground",
-      warning: "text-warning-900",
-      success: "text-success-900",
-      highlight: "text-highlight-900",
-      info: "text-info-900",
-      green: "text-success-900",
-      blue: "text-highlight-900",
-      rose: "text-warning-900",
-      golden: "text-info-900",
-      outline: "text-muted-foreground",
+      error: "text-red-800 dark:text-red-200",
+      success: "text-success-800",
+      info: "text-highlight-800",
+      warning: "text-orange-800 dark:text-orange-200",
+      gray: "text-gray-800 dark:text-gray-200",
     },
   },
 });
@@ -131,41 +68,52 @@ function ContentMessage({
   title,
   variant = "info",
   children,
-  size = "md",
-  className = "",
+  size = "sm",
+  className,
   icon,
   action,
 }: ContentMessageProps) {
+  const colorClass = colorVariants({ variant });
+  const iconSize = size === "sm" ? "sm" : "xs";
+  const titleClass = size === "sm" ? "heading-base" : "heading-sm";
+  const bodyClass = size === "sm" ? "copy-sm" : "copy-xs";
+  const innerGap = size === "sm" ? "gap-2" : "gap-1.5";
+
   return (
-    <div className={cn(contentMessageVariants({ variant, size }), className)}>
-      <div
-        className={cn(
-          "flex gap-3",
-          action ? "items-center justify-between" : "flex-col"
-        )}
-      >
-        <div className="flex flex-col gap-1">
-          {(icon || title) && (
-            <div className="flex items-center gap-1.5">
-              {icon && (
-                <Icon
-                  size="sm"
-                  visual={icon}
-                  className={iconVariants({ variant })}
-                />
-              )}
-              {title && (
-                <div className={titleVariants({ variant })}>{title}</div>
-              )}
-            </div>
-          )}
+    <div className={cn(containerVariants({ variant, size }), className)}>
+      {title ? (
+        <div className={cn("flex flex-col", innerGap)}>
+          <div className={cn("flex items-center", innerGap)}>
+            {icon && (
+              <Icon
+                size={iconSize}
+                visual={icon}
+                className={cn("shrink-0", colorClass)}
+              />
+            )}
+            <span className={cn(titleClass, colorClass)}>{title}</span>
+          </div>
           {children && (
-            <div className={textVariants({ variant })}>{children}</div>
+            <div className={cn(bodyClass, colorClass)}>{children}</div>
           )}
         </div>
-        {action && <div className="shrink-0">{action}</div>}
-      </div>
-      {/* TODO(2025-08-13 aubin): Allow passing a ContentMessageAction here. */}
+      ) : (
+        <div className={cn("flex items-start", innerGap)}>
+          {icon && (
+            <Icon
+              size={iconSize}
+              visual={icon}
+              className={cn("shrink-0 mt-px", colorClass)}
+            />
+          )}
+          {children && (
+            <div className={cn("flex-1 min-w-0", bodyClass, colorClass)}>
+              {children}
+            </div>
+          )}
+        </div>
+      )}
+      {action && <div className="shrink-0">{action}</div>}
     </div>
   );
 }
@@ -176,51 +124,4 @@ function ContentMessageAction(props: ButtonProps) {
   );
 }
 
-export interface ContentMessageInlineProps {
-  title?: string;
-  className?: string;
-  children?: React.ReactNode;
-  variant?: ContentMessageVariantType;
-  icon?: ComponentType;
-}
-
-function ContentMessageInline({
-  title,
-  variant = "info",
-  children,
-  className = "",
-  icon,
-}: ContentMessageInlineProps) {
-  const childrenArray = React.Children.toArray(children);
-
-  const { actionChilds, contentChildren } = childrenArray.reduce(
-    ({ actionChilds, contentChildren }, child) => {
-      if (React.isValidElement(child) && child.type === ContentMessageAction) {
-        actionChilds.push(child);
-      } else {
-        contentChildren.push(child);
-      }
-      return { actionChilds, contentChildren };
-    },
-    {
-      actionChilds: [] as React.ReactNode[],
-      contentChildren: [] as React.ReactNode[],
-    }
-  );
-
-  return (
-    <div className={cn(contentMessageInlineVariants({ variant }), className)}>
-      {icon && (
-        <Icon size="sm" visual={icon} className={iconVariants({ variant })} />
-      )}
-      <div className={cn("flex-1 min-w-0", textVariants({ variant }))}>
-        {title && <span className={titleVariants({ variant })}>{title}</span>}
-        {title && contentChildren.length > 0 && ": "}
-        {contentChildren}
-      </div>
-      {actionChilds}
-    </div>
-  );
-}
-
-export { ContentMessage, ContentMessageAction, ContentMessageInline };
+export { ContentMessage, ContentMessageAction, CONTENT_MESSAGE_VARIANTS };
