+++ sparkle/src/components/index.ts
@@ -66,7 +66,7 @@ export {
 export {
   ContentMessage,
   ContentMessageAction,
-  ContentMessageInline,
+  CONTENT_MESSAGE_VARIANTS,
 } from "./ContentMessage";
 export { ContextItem } from "./ContextItem";
 export type {
@@ -213,10 +213,6 @@ export {
   NavTabPillList,
   NavTabPillTrigger,
 } from "./NavTabPill";
-export {
-  NewContentMessage,
-  NewContentMessageAction,
-} from "./NewContentMessage";
 export type { NotificationType } from "./Notification";
 export { Notification, useSendNotification } from "./Notification";
 export { NotificationButton } from "./NotificationButton";
